refactor: bannerSlide를 ui 폴더로 이동하고 이미지 사이즈 조정
